+++ web/apps/photos/src/components/pages/gallery/Navbar.tsx
@@ -5,7 +5,7 @@ import { FlexWrapper, HorizontalFlex } from "@ente/shared/components/Container";
 import ArrowBack from "@mui/icons-material/ArrowBack";
 import MenuIcon from "@mui/icons-material/Menu";
 import { IconButton, Typography } from "@mui/material";
-import SearchBar from "components/Search/SearchBar";
+import SearchBar from "components/SearchBar";
 import UploadButton from "components/Upload/UploadButton";
 import { t } from "i18next";
 import { Collection } from "types/collection";
